[Behat] Revert moving both simple and configurable product pages into subdirectory
