+++ src/Sylius/Behat/Context/Ui/Admin/ManagingProductsContext.php
@@ -24,12 +24,12 @@ use Sylius\Behat\Element\Admin\Product\ProductTranslationsFormElementInterface;
 use Sylius\Behat\NotificationType;
 use Sylius\Behat\Page\Admin\Crud\CreatePageInterface;
 use Sylius\Behat\Page\Admin\Crud\UpdatePageInterface;
-use Sylius\Behat\Page\Admin\Product\ConfigurableProduct\CreateConfigurableProductPageInterface;
-use Sylius\Behat\Page\Admin\Product\ConfigurableProduct\UpdateConfigurableProductPageInterface;
+use Sylius\Behat\Page\Admin\Product\CreateConfigurableProductPageInterface;
+use Sylius\Behat\Page\Admin\Product\CreateSimpleProductPageInterface;
 use Sylius\Behat\Page\Admin\Product\IndexPageInterface;
 use Sylius\Behat\Page\Admin\Product\IndexPerTaxonPageInterface;
-use Sylius\Behat\Page\Admin\Product\SimpleProduct\CreateSimpleProductPageInterface;
-use Sylius\Behat\Page\Admin\Product\SimpleProduct\UpdateSimpleProductPageInterface;
+use Sylius\Behat\Page\Admin\Product\UpdateConfigurableProductPageInterface;
+use Sylius\Behat\Page\Admin\Product\UpdateSimpleProductPageInterface;
 use Sylius\Behat\Page\Admin\ProductReview\IndexPageInterface as ProductReviewIndexPageInterface;
 use Sylius\Behat\Page\Admin\ProductVariant\CreatePageInterface as VariantCreatePageInterface;
 use Sylius\Behat\Page\Admin\ProductVariant\GeneratePageInterface;
