+++ src/Sylius/Behat/Context/Ui/Admin/ProductCreationContext.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace Sylius\Behat\Context\Ui\Admin;
 
 use Behat\Behat\Context\Context;
-use Sylius\Behat\Page\Admin\Product\SimpleProduct\CreateSimpleProductPageInterface;
+use Sylius\Behat\Page\Admin\Product\CreateSimpleProductPageInterface;
 use Sylius\Component\Core\Formatter\StringInflector;
 use Sylius\Component\Core\Model\ChannelInterface;
 use Sylius\Component\Core\Model\TaxonInterface;
