+++ src/Sylius/Behat/Page/Admin/Product/CreateConfigurableProductPage.php
@@ -11,75 +11,32 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Behat\Page\Admin\Product\ConfigurableProduct;
+namespace Sylius\Behat\Page\Admin\Product;
 
 use Behat\Mink\Element\NodeElement;
 use Behat\Mink\Session;
-use Sylius\Behat\Behaviour\ChecksCodeImmutability;
-use Sylius\Behat\Page\Admin\Crud\UpdatePage as BaseUpdatePage;
-use Sylius\Behat\Service\AutocompleteHelper;
+use Sylius\Behat\Page\Admin\Crud\CreatePage as BaseCreatePage;
+use Sylius\Behat\Service\DriverHelper;
 use Sylius\Behat\Service\Helper\AutocompleteHelperInterface;
 use Symfony\Component\Routing\RouterInterface;
 
-class UpdateConfigurableProductPage extends BaseUpdatePage implements UpdateConfigurableProductPageInterface
+class CreateConfigurableProductPage extends BaseCreatePage implements CreateConfigurableProductPageInterface
 {
-    use ChecksCodeImmutability;
-
-    /**
-     * @param array<array-key, string> $minkParameters
-     */
     public function __construct(
         Session $session,
         $minkParameters,
         RouterInterface $router,
         string $routeName,
-        private AutocompleteHelperInterface $autocompleteHelper,
+        private readonly AutocompleteHelperInterface $autocompleteHelper,
     ) {
         parent::__construct($session, $minkParameters, $router, $routeName);
     }
 
-    public function saveChanges(): void
+    public function create(): void
     {
         $this->waitForFormUpdate();
 
-        parent::saveChanges();
-    }
-
-    public function isProductOptionChosen(string $option): bool
-    {
-        $optionElement = $this->getElement('options')->getParent();
-
-        return AutocompleteHelper::isValueVisible($this->getSession(), $optionElement, $option);
-    }
-
-    public function isProductOptionsDisabled(): bool
-    {
-        return 'disabled' === $this->getElement('options')->getAttribute('disabled');
-    }
-
-    public function hasTab(string $name): bool
-    {
-        return $this->hasElement('side_navigation_tab', ['%name%' => $name]);
-    }
-
-    public function checkChannel(string $channelCode): void
-    {
-        $this->getElement('channel', ['%channel_code%' => $channelCode])->check();
-    }
-
-    public function goToVariantsList(): void
-    {
-        $this->getDocument()->clickLink('List variants');
-    }
-
-    public function goToVariantCreation(): void
-    {
-        $this->getDocument()->clickLink('Create');
-    }
-
-    public function goToVariantGeneration(): void
-    {
-        $this->getDocument()->clickLink('Generate');
+        parent::create();
     }
 
     public function specifyCode(string $code): void
@@ -94,15 +51,10 @@ class UpdateConfigurableProductPage extends BaseUpdatePage implements UpdateConf
 
     public function selectOption(string $optionName): void
     {
-        $this->changeTab('details');
+        $this->changeTab();
         $productOptionsAutocomplete = $this->getElement('product_options_autocomplete');
 
         $this->autocompleteHelper->selectByName($this->getDriver(), $productOptionsAutocomplete->getXpath(), $optionName);
-    }
-
-    protected function getCodeElement(): NodeElement
-    {
-        return $this->getElement('code');
     }
 
     /**
@@ -119,7 +71,6 @@ class UpdateConfigurableProductPage extends BaseUpdatePage implements UpdateConf
                 'code' => '[data-test-code]',
                 'enabled' => '[data-test-enabled]',
                 'form' => '[data-live-name-value="sylius_admin:product:form"]',
-                'options' => '[data-test-options]',
                 'product_options_autocomplete' => '[data-test-product-options-autocomplete]',
                 'side_navigation_tab' => '[data-test-side-navigation-tab="%name%"]',
             ],
@@ -142,5 +93,14 @@ class UpdateConfigurableProductPage extends BaseUpdatePage implements UpdateConf
         $form->waitFor(1500, function () use ($form) {
             return !$form->hasAttribute('busy');
         });
+    }
+
+    private function changeTab(): void
+    {
+        if (DriverHelper::isNotJavascript($this->getDriver())) {
+            return;
+        }
+
+        $this->getElement('side_navigation_tab', ['%name%' => 'details'])->click();
     }
 }
