+++ src/Sylius/Behat/Page/Admin/Product/UpdateConfigurableProductPageInterface.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Behat\Page\Admin\Product\ConfigurableProduct;
+namespace Sylius\Behat\Page\Admin\Product;
 
 use Sylius\Behat\Page\Admin\Crud\UpdatePageInterface;
