+++ src/Sylius/Behat/Resources/config/services/pages/admin/product.xml
@@ -13,13 +13,13 @@
 
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <parameters>
-        <parameter key="sylius.behat.page.admin.product.create_configurable.class">Sylius\Behat\Page\Admin\Product\ConfigurableProduct\CreateConfigurableProductPage</parameter>
-        <parameter key="sylius.behat.page.admin.product.create_simple.class">Sylius\Behat\Page\Admin\Product\SimpleProduct\CreateSimpleProductPage</parameter>
+        <parameter key="sylius.behat.page.admin.product.create_configurable.class">Sylius\Behat\Page\Admin\Product\CreateConfigurableProductPage</parameter>
+        <parameter key="sylius.behat.page.admin.product.create_simple.class">Sylius\Behat\Page\Admin\Product\CreateSimpleProductPage</parameter>
         <parameter key="sylius.behat.page.admin.product.index.class">Sylius\Behat\Page\Admin\Product\IndexPage</parameter>
         <parameter key="sylius.behat.page.admin.product.index_per_taxon.class">Sylius\Behat\Page\Admin\Product\IndexPerTaxonPage</parameter>
         <parameter key="sylius.behat.page.admin.product.show.class">Sylius\Behat\Page\Admin\Product\ShowPage</parameter>
-        <parameter key="sylius.behat.page.admin.product.update_simple.class">Sylius\Behat\Page\Admin\Product\SimpleProduct\UpdateSimpleProductPage</parameter>
-        <parameter key="sylius.behat.page.admin.product.update_configurable.class">Sylius\Behat\Page\Admin\Product\ConfigurableProduct\UpdateConfigurableProductPage</parameter>
+        <parameter key="sylius.behat.page.admin.product.update_simple.class">Sylius\Behat\Page\Admin\Product\UpdateSimpleProductPage</parameter>
+        <parameter key="sylius.behat.page.admin.product.update_configurable.class">Sylius\Behat\Page\Admin\Product\UpdateConfigurableProductPage</parameter>
     </parameters>
 
     <services>
